modulo_03_gcp/pyproject.toml
[project]
name = "modulo-03-gcp"
version = "0.1.0"
description = "Módulo 03: Vertex AI — RAG, LangChain y Agentes con LLMs en Google Cloud"
requires-python = ">=3.11"
dependencies = [
    # ── Google Generative AI (Gemini / Vertex AI unified SDK) ──────────────────
    "google-genai>=1.15.0",
    "google-cloud-aiplatform>=1.97.0",

    # ── LangChain core + Vertex AI integration ─────────────────────────────────
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",

    # ── LangGraph para agentes ─────────────────────────────────────────────────
    "langgraph>=0.2.0",

    # ── RAG: vector stores + embeddings ───────────────────────────────────────
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",

    # ── Utilidades ─────────────────────────────────────────────────────────────
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pypdf>=4.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Instala los 3 paquetes del módulo como top-level (sin sys.path hacks)
# Nota: Python no permite identificadores que empiecen con dígito.
# Las carpetas están numeradas (01_*, 02_*, 03_*) para la vista del filesystem,
# pero se instalan con nombres Python válidos mediante launchers en la raíz.
packages = ["src/shared"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipykernel>=6.0.0",
    "jupyterlab",
    "ruff>=0.4.0",
]
